This introduces a change to config.go to support the use of .env file that holds the needed environmental variables and their values
A new function, getEnvFilePath, has been written that returns the full folder path to where the mcp binary is located as this is where we expect the .env file to be.

getEnvFilePath uses filepath.Abs(os.Args[0]) which to obtain full folder path and the binary filename which is then split with filepath.Split to obtain
the full folder path which is then returned.