test: add integration test helpers for easily adding new tests
Tests will use testcontainers where we spin up a Neo4j database and run
our integration tests against them.

As there was a deliberate choice to use only one container (as opposed
to a container for each test), there's a need to isolate tests, which we
do by adding test IDs to data added to the database. This way each test
is responsible for its own data creation and cleanup.

Along with this there are various Neo4j data structure specific assertion
helper functions added. This is quite opinionated, but the aim is to make
it easy for contributors to add new tests.

Note that waitForConnectibityNoT is a custom function which waits for
Neo4j to be ready to accept connections using exponential backoff retry
logic. It's used during test setup to ensure the database container is
fully initialized before running tests.